Moved VariableNotAssignedInspectionTests for QuickFixes into relevant files under RD.QuickFixes
