+++ RubberduckTests/Inspections/VariableNotAssignedInspectionTests.cs
@@ -1,7 +1,6 @@
 using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Rubberduck.Inspections.Concrete;
-using Rubberduck.Inspections.QuickFixes;
 using Rubberduck.Parsing.Inspections.Resources;
 using Rubberduck.VBEditor.SafeComWrappers.Abstract;
 using RubberduckTests.Mocks;
@@ -110,132 +109,6 @@ End Sub";
             var inspectionResults = inspection.GetInspectionResults();
 
             Assert.IsFalse(inspectionResults.Any());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void UnassignedVariable_QuickFixWorks()
-        {
-            const string inputCode =
-@"Sub Foo()
-Dim var1 as Integer
-End Sub";
-
-            const string expectedCode =
-@"Sub Foo()
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new VariableNotAssignedInspection(state);
-            new RemoveUnassignedIdentifierQuickFix(state).Fix(inspection.GetInspectionResults().First());
-
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void UnassignedVariable_VariableOnMultipleLines_QuickFixWorks()
-        {
-            const string inputCode =
-@"Sub Foo()
-Dim _
-var1 _
-as _
-Integer
-End Sub";
-
-            const string expectedCode =
-@"Sub Foo()
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new VariableNotAssignedInspection(state);
-            new RemoveUnassignedIdentifierQuickFix(state).Fix(inspection.GetInspectionResults().First());
-
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void UnassignedVariable_MultipleVariablesOnSingleLine_QuickFixWorks()
-        {
-            const string inputCode =
-@"Sub Foo()
-Dim var1 As Integer, var2 As Boolean
-End Sub";
-
-            // note the extra space after "Integer"--the VBE will remove it
-            const string expectedCode =
-@"Sub Foo()
-Dim var1 As Integer
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new VariableNotAssignedInspection(state);
-            new RemoveUnassignedIdentifierQuickFix(state).Fix(
-                inspection.GetInspectionResults().Single(s => s.Target.IdentifierName == "var2"));
-
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void UnassignedVariable_MultipleVariablesOnMultipleLines_QuickFixWorks()
-        {
-            const string inputCode =
-@"Sub Foo()
-Dim var1 As Integer, _
-var2 As Boolean
-End Sub";
-
-            const string expectedCode =
-@"Sub Foo()
-Dim var1 As Integer
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new VariableNotAssignedInspection(state);
-            new RemoveUnassignedIdentifierQuickFix(state).Fix(
-                inspection.GetInspectionResults().Single(s => s.Target.IdentifierName == "var2"));
-
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void UnassignedVariable_IgnoreQuickFixWorks()
-        {
-            const string inputCode =
-@"Sub Foo()
-Dim var1 as Integer
-End Sub";
-
-            const string expectedCode =
-@"Sub Foo()
-'@Ignore VariableNotAssigned
-Dim var1 as Integer
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new VariableNotAssignedInspection(state);
-            new IgnoreOnceQuickFix(state, new[] {inspection}).Fix(inspection.GetInspectionResults().First());
-
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
 
         [TestMethod]
